Rename list related files
