+++ packages/devextreme/js/__internal/scheduler/tooltip_strategies/m_tooltip_strategy_base.ts
@@ -3,7 +3,7 @@ import { FunctionTemplate } from '@js/core/templates/function_template';
 import { isRenderer } from '@js/core/utils/type';
 import Button from '@js/ui/button';
 import { createPromise } from '@ts/core/utils/promise';
-import List from '@ts/ui/list/m_list.edit';
+import List from '@ts/ui/list/list.edit';
 
 const TOOLTIP_APPOINTMENT_ITEM = 'dx-tooltip-appointment-item';
 const TOOLTIP_APPOINTMENT_ITEM_CONTENT = `${TOOLTIP_APPOINTMENT_ITEM}-content`;
