+++ packages/devextreme/js/__internal/ui/drop_down_editor/m_drop_down_list.ts
@@ -28,8 +28,8 @@ import type { Properties as PopupProperties } from '@js/ui/popup';
 import errors from '@js/ui/widget/ui.errors';
 import type { OptionChanged } from '@ts/core/widget/types';
 import DropDownEditor from '@ts/ui/drop_down_editor/m_drop_down_editor';
-import type { ListBaseProperties } from '@ts/ui/list/m_list.base';
-import List from '@ts/ui/list/m_list.edit.search';
+import type { ListBaseProperties } from '@ts/ui/list/list.base';
+import List from '@ts/ui/list/list.edit.search';
 import DataConverterMixin from '@ts/ui/shared/m_grouped_data_converter_mixin';
 
 const window = getWindow();
