+++ packages/devextreme/js/__internal/ui/hierarchical_collection/data_adapter.ts
@@ -6,7 +6,7 @@ import { extend } from '@js/core/utils/extend';
 import { each } from '@js/core/utils/iterator';
 import { isDefined, isFunction } from '@js/core/utils/type';
 import errors from '@js/ui/widget/ui.errors';
-import SearchBoxController, { getOperationBySearchMode } from '@ts/ui/collection/m_search_box_mixin';
+import SearchBoxController, { getOperationBySearchMode } from '@ts/ui/collection/search_box_controller';
 import TextBox from '@ts/ui/text_box/m_text_box';
 
 import HierarchicalDataConverter, {
