+++ packages/devextreme/js/__internal/ui/list/modules/context.ts
@@ -1,0 +1,1 @@
+import '../list.edit.decorator.context';
