+++ packages/devextreme/js/__internal/ui/list/modules/deleting.context.ts
@@ -1,0 +1,1 @@
+import './context';
