+++ packages/devextreme/js/__internal/ui/list/modules/deleting.slideItem.ts
@@ -1,0 +1,1 @@
+import '../list.edit.decorator.switchable.slide';
