+++ packages/devextreme/js/__internal/ui/list/modules/deleting.static.ts
@@ -1,0 +1,1 @@
+import '../list.edit.decorator.static';
