+++ packages/devextreme/js/__internal/ui/list/modules/deleting.swipe.ts
@@ -1,0 +1,1 @@
+import '../list.edit.decorator.swipe';
