+++ packages/devextreme/js/__internal/ui/list/modules/deleting.ts
@@ -1,0 +1,6 @@
+import './deleting.context';
+import './deleting.slideButton';
+import './deleting.slideItem';
+import './deleting.static';
+import './deleting.swipe';
+import './deleting.toggle';
