+++ packages/devextreme/js/__internal/ui/list/modules/dragging.ts
@@ -1,0 +1,1 @@
+import '../list.edit.decorator.reorder';
