--- packages/devextreme/js/__internal/ui/list/modules/m_deleting.context.ts
@@ -1,1 +1,0 @@
-import './m_context';
