--- packages/devextreme/js/__internal/ui/list/modules/m_deleting.static.ts
@@ -1,1 +1,0 @@
-import '../m_list.edit.decorator.static';
