--- packages/devextreme/js/__internal/ui/list/modules/m_deleting.toggle.ts
@@ -1,1 +1,0 @@
-import '../m_list.edit.decorator.switchable.button';
