--- packages/devextreme/js/__internal/ui/list/modules/m_deleting.ts
@@ -1,6 +1,0 @@
-import './m_deleting.context';
-import './m_deleting.slideButton';
-import './m_deleting.slideItem';
-import './m_deleting.static';
-import './m_deleting.swipe';
-import './m_deleting.toggle';
