--- packages/devextreme/js/__internal/ui/list/modules/m_selection.ts
@@ -1,1 +1,0 @@
-import '../m_list.edit.decorator.selection';
