+++ packages/devextreme/js/__internal/ui/list/modules/search.ts
@@ -1,4 +1,4 @@
-import SearchBoxController from '@ts/ui/collection/m_search_box_mixin';
+import SearchBoxController from '@ts/ui/collection/search_box_controller';
 import TextBox from '@ts/ui/text_box/m_text_box';
 
 SearchBoxController.setEditorClass(TextBox);
