+++ packages/devextreme/js/__internal/ui/list/modules/selection.ts
@@ -1,0 +1,1 @@
+import '../list.edit.decorator.selection';
