+++ packages/devextreme/js/__internal/ui/tree_view/tree_view.search.ts
@@ -3,8 +3,8 @@ import type { dxElementWrapper } from '@js/core/renderer';
 import $ from '@js/core/renderer';
 import type { DxEvent } from '@js/events';
 import type { OptionChanged } from '@ts/core/widget/types';
-import type { SearchBoxControllerOptions } from '@ts/ui/collection/m_search_box_mixin';
-import SearchBoxController from '@ts/ui/collection/m_search_box_mixin';
+import type { SearchBoxControllerOptions } from '@ts/ui/collection/search_box_controller';
+import SearchBoxController from '@ts/ui/collection/search_box_controller';
 import type { DataAdapterOptions } from '@ts/ui/hierarchical_collection/data_adapter';
 import type { InternalNode } from '@ts/ui/hierarchical_collection/data_converter';
 import TextBox from '@ts/ui/text_box/m_text_box';
