+++ packages/devextreme/js/ui/list.js
@@ -1,9 +1,9 @@
 import List from './list_light';
-import '../__internal/ui/list/modules/m_selection';
-import '../__internal/ui/list/modules/m_deleting';
-import '../__internal/ui/list/modules/m_dragging';
-import '../__internal/ui/list/modules/m_context';
-import '../__internal/ui/list/modules/m_search';
+import '../__internal/ui/list/modules/selection';
+import '../__internal/ui/list/modules/deleting';
+import '../__internal/ui/list/modules/dragging';
+import '../__internal/ui/list/modules/context';
+import '../__internal/ui/list/modules/search';
 
 export default List;
