+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.scheduler/desktopTooltip.tests.js
@@ -2,7 +2,7 @@ import { DesktopTooltipStrategy } from '__internal/scheduler/tooltip_strategies/
 import { FunctionTemplate } from 'core/templates/function_template';
 import { extend } from 'core/utils/extend';
 import Tooltip from 'ui/tooltip';
-import List from '__internal/ui/list/m_list.edit';
+import List from '__internal/ui/list/list.edit';
 import Button from 'ui/button';
 import support from '__internal/core/utils/m_support';
