+++ packages/devextreme/testing/tests/DevExpress.ui.widgets/listParts/editingUITests.js
@@ -9,9 +9,9 @@ import config from 'core/config';
 import pointerMock from '../../../helpers/pointerMock.js';
 import contextMenuEvent from 'common/core/events/contextmenu';
 import keyboardMock from '../../../helpers/keyboardMock.js';
-import { registry, register } from '__internal/ui/list/m_list.edit.decorator_registry';
-import SwitchableEditDecorator from '__internal/ui/list/m_list.edit.decorator.switchable';
-import SwitchableButtonEditDecorator from '__internal/ui/list/m_list.edit.decorator.switchable.button';
+import { registry, register } from '__internal/ui/list/list.edit.decorator_registry';
+import SwitchableEditDecorator from '__internal/ui/list/list.edit.decorator.switchable';
+import SwitchableButtonEditDecorator from '__internal/ui/list/list.edit.decorator.switchable.button';
 import themes from 'ui/themes';
 import { DataSource } from 'common/data/data_source/data_source';
 import ArrayStore from 'common/data/array_store';
